remove newVertex/newEdge/newVertexId on S2GraphLike.
