+++ s2core/src/test/scala/org/apache/s2graph/core/tinkerpop/process/S2GraphProcessStandardTest.scala
@@ -19,7 +19,7 @@
 
 package org.apache.s2graph.core.tinkerpop.process
 
-import org.apache.s2graph.core.S2Graph
+import org.apache.s2graph.core.{S2Graph, S2GraphTp}
 import org.apache.s2graph.core.tinkerpop.S2GraphProvider
 import org.apache.tinkerpop.gremlin.GraphProviderClass
 import org.apache.tinkerpop.gremlin.process.ProcessStandardSuite
@@ -29,7 +29,7 @@ import org.junit.runners.MethodSorters
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 @RunWith(classOf[ProcessStandardSuite])
-@GraphProviderClass(provider = classOf[S2GraphProvider], graph = classOf[S2Graph])
+@GraphProviderClass(provider = classOf[S2GraphProvider], graph = classOf[S2GraphTp])
 class S2GraphProcessStandardTest {
 
 }
