+++ s2core/src/test/scala/org/apache/s2graph/core/tinkerpop/structure/S2GraphStructureStandardTest.scala
@@ -19,7 +19,7 @@
 
 package org.apache.s2graph.core.tinkerpop.structure
 
-import org.apache.s2graph.core.S2Graph
+import org.apache.s2graph.core.{S2Graph, S2GraphTp}
 import org.apache.s2graph.core.tinkerpop.S2GraphProvider
 import org.apache.tinkerpop.gremlin.structure.StructureStandardSuite
 import org.apache.tinkerpop.gremlin.{GraphProviderClass, LoadGraphWith}
@@ -29,7 +29,7 @@ import org.junit.runners.MethodSorters
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 @RunWith(classOf[StructureStandardSuite])
-@GraphProviderClass(provider = classOf[S2GraphProvider], graph = classOf[S2Graph])
+@GraphProviderClass(provider = classOf[S2GraphProvider], graph = classOf[S2GraphTp])
 class S2GraphStructureStandardTest {
 
 }
